[Refactor] recoil interface 이동, refMap 타입 수정
