+++ utils/recoil/application.ts
@@ -1,13 +1,9 @@
 import { atom } from 'recoil';
 import { v1 } from 'uuid';
-import { AlertColor } from '@mui/material';
-import { IApplicantAnswer } from 'types/recruit/recruitments';
-
-export interface IapplicationAlertState {
-  alertState: boolean;
-  message: string;
-  severity: AlertColor | undefined;
-}
+import {
+  IapplicationAlertState,
+  IapplicationModal,
+} from 'types/recruit/recruitments';
 
 export const applicationAlertState = atom<IapplicationAlertState>({
   key: `applicationAlertState/${v1()}`,
@@ -17,12 +13,6 @@ export const applicationAlertState = atom<IapplicationAlertState>({
     severity: undefined,
   },
 });
-
-export interface IapplicationModal {
-  state: boolean;
-  content: 'APPLY' | 'CANCEL' | 'EDIT' | 'NONE';
-  formData: IApplicantAnswer[];
-}
 
 export const applicationModalState = atom<IapplicationModal>({
   key: `applicationModalState/${v1()}`,
